don't trust `totalItems` but just iterate over `orderedItems`
In one of my archives, `totalItems` in `outbox.json` was over 1,000 items short. I've fixed this by iterating directly on `orderedItems`.

Counting `orderedItems` first with `len` made the script pause a good amount of time, so I switched tqdm to dynamic mode in the interest of speed.

(It's interesting to note that [the spec for `totalItems`](https://www.w3.org/TR/activitystreams-vocabulary/#dfn-totalitems) says "This number might not reflect the actual number of items serialized…".)